refactor(imports): allow importing from hooks dir
