+++ src/pages/JobEdit/JobEdit.js
@@ -2,7 +2,7 @@ import React, { useState } from 'react'
 import { Card, IconInfo16, NoticeBox } from '@dhis2/ui'
 import { useParams } from 'react-router-dom'
 import i18n from '@dhis2/d2-i18n'
-import { useJobById } from '../../hooks/jobs'
+import { useJobById } from '../../hooks'
 import { DiscardFormButton } from '../../components/Buttons'
 import { JobEditFormContainer } from '../../components/Forms'
 import { JobDetails } from '../../components/JobDetails'
